+++ code/src/cljs/sixsq/nuvla/ui/edge_detail/subs.cljs
@@ -163,3 +163,27 @@
   ::nuvlabox-current-playbook
   (fn [db]
     (::spec/nuvlabox-current-playbook db)))
+
+
+(reg-sub
+  ::nuvlabox-log
+  (fn [db]
+    (::spec/nuvlabox-log db)))
+
+
+(reg-sub
+  ::nuvlabox-log-id
+  (fn [db]
+    (::spec/nuvlabox-log-id db)))
+
+
+(reg-sub
+  ::nuvlabox-log-since
+  (fn [db]
+    (::spec/nuvlabox-log-since db)))
+
+
+(reg-sub
+  ::nuvlabox-log-play?
+  (fn [db]
+    (::spec/nuvlabox-log-play? db)))